Finish Spanning Tree

+++ ThuPVM15_PF3_Final_Project/src/Services/Geometry_Services.java
@@ -5,7 +5,7 @@ import Models.Node;
 
 public class Geometry_Services {
 
-	public static double distance_between_2_points(int x1, int y1, int x2, int y2) {
+	public static double distance_between_2_points(double x1, double y1, double x2, double y2) {
 		return Math.sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
 	}
